+++ business/database/02_schema.sql
@@ -2184,7 +2184,8 @@ ALTER TABLE `object_relation__object_relation`
 -- Contraintes pour la table `obtained_units`
 --
 ALTER TABLE `obtained_units`
-  ADD CONSTRAINT `obtained_units_ibfk_1` FOREIGN KEY (`first_deployment_mission`) REFERENCES `missions` (`id`);
+  ADD CONSTRAINT `obtained_units_ibfk_1` FOREIGN KEY (`first_deployment_mission`) REFERENCES `missions` (`id`),
+  ADD CONSTRAINT `obtained_units_mission_fk` FOREIGN KEY (`mission_id`) REFERENCES `missions` (`id`) ON DELETE SET NULL;
 
 --
 -- Contraintes pour la table `obtained_upgrades`